Added `logout` button and logout page
Made a logged_out_required decorator which restricts the login page
(unless you are not logged in). I also made the `login` button in the
nav bar change to `logout` when you login in to the site and created a
logout page that confirms that the user wants to log out.